ºC to F (int to float)

+++ exercises/temperature-converter.py
@@ -1,3 +1,3 @@
-celsius = int(input("Insert the temperature in ºC to be converter in Fahrenheit: "))
+celsius = float(input("Insert the temperature in ºC to be converter in Fahrenheit: "))
 fh = (celsius*9/5) + 32
 print(f"It will be in {fh:.2f}ºF ")
